Organize SyntaxWalker classes
