+++ RefactorMCP.ConsoleApp/SyntaxRewriters/CalledMethodCollector.cs
@@ -1,3 +1,4 @@
+using RefactorMCP.ConsoleApp.SyntaxWalkers;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Generic;
