--- RefactorMCP.ConsoleApp/SyntaxRewriters/NameCollectorWalker.cs
@@ -1,10 +1,0 @@
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Collections.Generic;
-
-internal abstract class NameCollectorWalker : CSharpSyntaxWalker
-{
-    public HashSet<string> Names { get; } = new();
-
-    protected void Add(string name) => Names.Add(name);
-}
