--- RefactorMCP.ConsoleApp/SyntaxRewriters/PrivateFieldUsageWalker.cs
@@ -1,12 +1,0 @@
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System.Collections.Generic;
-
-internal class PrivateFieldUsageWalker : TrackedNameWalker
-{
-    public HashSet<string> UsedFields => Matches;
-
-    public PrivateFieldUsageWalker(HashSet<string> privateFieldNames)
-        : base(privateFieldNames)
-    {
-    }
-}
