--- RefactorMCP.ConsoleApp/SyntaxRewriters/StaticFieldNameWalker.cs
@@ -1,16 +1,0 @@
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-
-internal class StaticFieldNameWalker : NameCollectorWalker
-{
-    public override void VisitFieldDeclaration(FieldDeclarationSyntax node)
-    {
-        if (node.Modifiers.Any(SyntaxKind.StaticKeyword))
-        {
-            foreach (var variable in node.Declaration.Variables)
-                Add(variable.Identifier.ValueText);
-        }
-        base.VisitFieldDeclaration(node);
-    }
-}
