+++ RefactorMCP.ConsoleApp/SyntaxWalkers/ClassCollectorWalker.cs
@@ -1,7 +1,7 @@
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Generic;
 
-namespace RefactorMCP.ConsoleApp.SyntaxRewriters
+namespace RefactorMCP.ConsoleApp.SyntaxWalkers
 {
     internal class ClassCollectorWalker : TypeCollectorWalker<ClassDeclarationSyntax>
     {
