+++ RefactorMCP.ConsoleApp/SyntaxWalkers/InterfaceCollectorWalker.cs
@@ -1,0 +1,10 @@
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using System.Collections.Generic;
+namespace RefactorMCP.ConsoleApp.SyntaxWalkers
+{
+
+    internal class InterfaceCollectorWalker : TypeCollectorWalker<InterfaceDeclarationSyntax>
+    {
+        public Dictionary<string, InterfaceDeclarationSyntax> Interfaces => Types;
+    }
+}
