+++ RefactorMCP.ConsoleApp/SyntaxWalkers/MethodMetricsWalker.cs
@@ -4,7 +4,7 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace RefactorMCP.ConsoleApp.SyntaxRewriters
+namespace RefactorMCP.ConsoleApp.SyntaxWalkers
 {
     internal class MethodMetricsWalker : CSharpSyntaxWalker
     {
