+++ RefactorMCP.ConsoleApp/SyntaxWalkers/RefactoringOpportunityWalker.cs
@@ -2,7 +2,7 @@ using Microsoft.CodeAnalysis;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace RefactorMCP.ConsoleApp.SyntaxRewriters
+namespace RefactorMCP.ConsoleApp.SyntaxWalkers
 {
     internal class RefactoringOpportunityWalker
     {
